// src/shared/types/db.ts
export namespace DB {
// ...
	export type ChuniProfileData = {
		id: number
		user: number
		version: number
		exp: number | null
		level: number | null
		point: number | null
		frameId: number | null
		isMaimai: boolean | null
		trophyId: number | null
		userName: string | null
		isWebJoin: boolean | null
		playCount: number | null
		lastGameId: string | null
		totalPoint: number | null
		characterId: number | null
		firstGameId: string | null
		friendCount: number | null
		lastPlaceId: number | null
		nameplateId: number | null
		totalMapNum: number | null
		lastAllNetId: number | null
		lastClientId: string | null
		lastPlayDate: string | null
		lastRegionId: number | null
		playerRating: number | null
		totalHiScore: number | null
		webLimitDate: string | null
		firstPlayDate: string | null
		highestRating: number | null
		lastPlaceName: string | null
		multiWinCount: number | null
		acceptResCount: number | null
		lastRegionName: string | null
		lastRomVersion: string | null
		multiPlayCount: number | null
		firstRomVersion: string | null
		lastDataVersion: string | null
		requestResCount: number | null
		successResCount: number | null
		eventWatchedDate: string | null
		firstDataVersion: string | null
		reincarnationNum: number | null
		playedTutorialBit: number | null
		totalBasicHighScore: number | null
		totalExpertHighScore: number | null
		totalMasterHighScore: number | null
		totalRepertoireCount: number | null
		firstTutorialCancelNum: number | null
		totalAdvancedHighScore: number | null
		masterTutorialCancelNum: number | null
		ext1: number | null
		ext2: number | null
		ext3: number | null
		ext4: number | null
		ext5: number | null
		ext6: number | null
		ext7: number | null
		ext8: number | null
		ext9: number | null
		ext10: number | null
		extStr1: string | null
		extStr2: string | null
		extLong1: number | null
		extLong2: number | null
		mapIconId: number | null
		compatibleCmVersion: string | null
		medal: number | null
		voiceId: number | null
		teamId: number | null
		eliteRankPoint: number
		stockedGridCount: number
		netBattleLoseCount: number
		netBattleHostErrCnt: number
		netBattle4thCount: number
		overPowerRate: number
		battleRewardStatus: number
		netBattle1stCount: number
		charaIllustId: number
		userNameEx: string
		netBattleWinCount: number
		netBattleCorrection: number
		classEmblemMedal: number
		overPowerPoint: number
		netBattleErrCnt: number
		battleRankId: number
		netBattle3rdCount: number
		netBattleConsecutiveWinCount: number
		overPowerLowerRank: number
		classEmblemBase: number
		battleRankPoint: number
		netBattle2ndCount: number
		totalUltimaHighScore: number
		skillId: number
		lastCountryCode: string
		isNetBattleHost: boolean
		battleRewardCount: number
		battleRewardIndex: number
		netBattlePlayCount: number
		exMapLoopCount: number
		netBattleEndState: number
		rankUpChallengeResults: string | null
		avatarBack: number
		avatarFace: number
		avatarPoint: number
		avatarItem: number
		avatarWear: number
		avatarFront: number
		avatarSkin: number
		avatarHead: number
		trophyIdSub1: number | null
		trophyIdSub2: number | null
	}
// ...
	export type ChuniProfileRating = {
		id: number
		user: number
		version: number
		type: string
		index: number
		musicId: number | null
		difficultId: number | null
		romVersionCode: number | null
		score: number | null
	}
// ...
}